refactoring: Extract coreos-cloud-init userdata validation
Extract it out of the config package into its own package "userdatavalidation" for future use in the upcoming node pools feature.
The former `coreosutil` package is now named `coreos/amiregistry` and the validation is in the `coreos/userdatavalidation` package to reflect the fact that both of them are coreos specific but they have different responsiblity compared to each other.
